impl missing Democracy precompile functions (#1840)

+++ precompiles/pallet-democracy/DemocracyInterface.sol
@@ -47,7 +47,7 @@ interface Democracy {
     function lowestUnbaked() external view returns (uint256);
 
     /// Get the details about an ongoing referendum.
-    /// @custom:selector e5a18359
+    /// @custom:selector f033b7cd
     ///
     /// @dev This, along with `finishedReferendumInfo`, wraps the pallet's `referendumInfo`
     /// function. It is necessary to split it into two here because Solidity only has c-style enums.
@@ -60,13 +60,13 @@ interface Democracy {
     ///  * The total aye vote (including conviction)
     ///  * The total nay vote (including conviction)
     ///  * The total turnout (not including conviction)
-    function ongoingReferendumInfo(uint256 refIndex)
+    function ongoingReferendumInfo(uint32 refIndex)
         external
         view
         returns (
             uint256,
             bytes32,
-            uint256,
+            uint8,
             uint256,
             uint256,
             uint256,
@@ -74,13 +74,13 @@ interface Democracy {
         );
 
     /// Get the details about a finished referendum.
-    /// @custom:selector 07df495b
+    /// @custom:selector c75abcce
     ///
     /// @dev This, along with `ongoingReferendumInfo`, wraps the pallet's `referendumInfo`
     /// function. It is necessary to split it into two here because Solidity only has c-style enums.
     /// @param refIndex The index of the referendum you are interested in
     /// @return A tuple including whether the referendum passed, and the block at which it finished.
-    function finishedReferendumInfo(uint256 refIndex)
+    function finishedReferendumInfo(uint32 refIndex)
         external
         view
         returns (bool, uint256);
